added cloud-api, have issues on displaying devices on app: fixing requried

+++ firmware/ezlopi/ezlopi-core/ezlopi_cloud/ezlopi_cloud.h
@@ -8,17 +8,17 @@ typedef struct s_ezlopi_cloud_info
 {
     // cloud information
     char device_name[32]; // device name, 'name' field in 'hub.devices.list'
-    uint16_t device_id;   //
+    uint32_t device_id;   //
     char *category;       // ezlopi-cloud/constants/category_str.h
     char *subcategory;    // ezlopi-cloud/constants/subcategory_str.h
     char *device_type;    // ezlopi-cloud/constants/device_types_str.h
     char *value_type;     // ezlopi-cloud/constants/values_str.h
     bool reachable;
     bool battery_powered;
-    uint16_t room_id;   //
+    uint32_t room_id;   //
     char room_name[32]; //
 
-    uint16_t item_id; //
+    uint32_t item_id; //
     char *item_name;  // https://api.ezlo.com/devices/items/index.html
 
     bool has_getter;
